🪁 :: move benchmarks.py
